+++ Modules/options.lua
@@ -105,6 +105,17 @@ function addon:OptionsTable()
 											addon:CallModule("version")
 										end,
 									},
+									sync = {
+										order = 10,
+										name = L["Sync"],
+										desc = L["Opens the synchronizer"],
+										type = "execute",
+										func = function()
+											InterfaceOptionsFrame:Hide()
+											LibStub("AceConfigDialog-3.0"):CloseAll()
+											self.Sync:Spawn()
+										end,
+									},
 								},
 							},
 							autoPassOptions = {
@@ -169,7 +180,7 @@ function addon:OptionsTable()
 										func = function() self:CallModule("history");	InterfaceOptionsFrame:Hide();end,
 									},
 									clearLootDB = {
-										order = -1,
+										order = 6,
 										name = L["Clear Loot History"],
 										desc = L["clear_loot_history_desc"],
 										type = "execute",
@@ -520,6 +531,12 @@ function addon:OptionsTable()
 										desc = L["observe_desc"],
 										type = "toggle",
 									},
+									autoAddRolls = {
+										order = 8,
+										name = L["Add Rolls"],
+										desc = L["add_rolls_desc"],
+										type = "toggle",
+									}
 								},
 							},
 							ignoreOptions = {
@@ -830,18 +847,18 @@ function addon:OptionsTable()
 							tierButtonsOptions = {
 								order = 2,
 									type = "group",
-									name = "Tier Buttons and Responses",
+									name = L["Tier Buttons and Responses"],
 									inline = true,
 									args = {
 										tierButtonsEnabled = {
 											order = 0,
-											name = "Enable Tier Buttons",
-											desc = "Check to enable a seperate set of buttons when rolling for tier pieces.",
+											name = L["Enable Tier Buttons"],
+											desc = L["enable_tierbuttons_desc"],
 											type = "toggle",
 										},
 										optionsDesc = {
 											order = 0.1,
-											name = "Select which buttons/responses to use when handling tier pieces. It's basically the same concept as above, except these are shown when tier pieces drop.\nA pass button is still added to the right.",
+											name = L["tier_buttons_desc"],
 											type = "description",
 											hidden = function() return not self.db.profile.tierButtonsEnabled end,
 										},